Fix race conditions by serializing code
Fixed earlier race conditions by serializing all of the code except for
the matrix-multiplication part.

Doing this at -O0 causes the the runtime for the basic correctness test
to go from 398.25s to 404.71s. While this does represent a slight
slowdown, a 1.5% increase in runtime seems like a good tradeoff for
correctness. The parallelism for the remaining code can be added back in
at a later time if needed.

Patch tested on my personal desktop and on the KNL nodes of Stampede.

Also modified CMakeLists.txt to have executables suffixed with the ".x"
suffix.